Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/900



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@6879 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/dtfmtsym.cpp
@@ -481,7 +481,7 @@ void
 DateFormatSymbols::initField(UnicodeString **field, int32_t& length, const UChar *data, LastResortSize numStr, LastResortSize strLen, UErrorCode &status) {
     if (U_SUCCESS(status)) {
         length = numStr;
-        *field = new UnicodeString[numStr];
+        *field = new UnicodeString[(size_t)numStr];
         if (*field) {
             for(int32_t i = 0; i<length; i++) {
                 *(*(field)+i) = data+(i*((int32_t)strLen));
